✨ 컴포넌트화 및 로그인 로직 수정
